changed to only look for explciit building and building keywords otherwise fallback to chat (follows chat-first)
