+++ src/modules/routing/index.ts
@@ -3,4 +3,3 @@ export * from "./schemas";
 export * from "./helpers";
 export * from "./router";
 export * from "./rules";
-export * from "./config";
